+++ test/container.spec.php
@@ -1,0 +1,9 @@
+<?php
+
+use Krak\Cargo;
+
+describe('Krak Cargo', function() {
+    describe('Container', function() {
+        require_once __DIR__ . '/container.php';
+    });
+});